requirements.txt
# AI-Powered Insurance Underwriting System
# Python 3.9+ required

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
pymongo>=4.6.0

# AI/ML - LangChain Core & Google
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.5
langgraph>=0.0.10
langchain-text-splitters>=0.0.1
# langchain-google-vertexai # Optional if user has project set up
faiss-cpu>=1.7.4

# AI/ML - Legacy (keeping for compatibility during migration)
fireworks-ai>=0.13.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Environment & Config
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.25.0
urllib3>=2.0.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
tenacity>=8.2.0
tiktoken>=0.5.2
